+++ lib/CHANGELOG.md
@@ -1,3 +1,7 @@
+# <unreleased>
+
+ * Updated crate description in toml file, overlooked in v1.1
+
 # 1.2.0 (October 27th, 2018)
 
  * Refactored and optimised the analysis engine