+++ lib/CHANGELOG.md
@@ -1,5 +1,6 @@
 # <unreleased>
 
+ * Updated usage docs per `OptionSetEx`/`OptionSet` pair introduction
  * Improved macro set construction testing
  * Updated crate description in toml file, overlooked in v1.1