+++ lib/CHANGELOG.md
@@ -1,4 +1,4 @@
-# <unreleased>
+# 1.2.1 (October 28th, 2018)
 
  * Updated usage docs per `OptionSetEx`/`OptionSet` pair introduction
  * Improved macro set construction testing